+++ packages/imperative/src/config/src/Config.ts
@@ -179,7 +179,9 @@ export class Config {
                 exists: false,
                 properties: Config.empty(),
                 global: layer === Layers.GlobalUser || layer === Layers.GlobalConfig,
-                user: layer === Layers.ProjectUser || layer === Layers.GlobalUser
+                user: layer === Layers.ProjectUser || layer === Layers.GlobalUser,
+                ignoreErrors: process.argv.includes("--help") || process.argv.includes("--version")
+                ||  process.argv[process.argv.length-1] === require.resolve('@zowe/cli')
             });
         }